removed code from common component to library export component
