+++ ui/src/app/navigation/libraryexport.component.html
@@ -1,0 +1,6 @@
+<button class="m-button" type="button" (click)=onDownloadLibrary() *ngIf="canExportLibrary">
+  <svg class="m-button-x-icon t-icon">
+    <use xlink:href="/assets/images/svg-defs.svg#icon-download"></use>
+  </svg>
+  <span class="m-button-x-text">Export RSD Library</span>
+</button>
